% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_calc_function.R
\name{age_calc}
\alias{age_calc}
\title{Calculating age from date of birth}
\usage{
age_calc(dob, enddate = Sys.Date(), units = "years", precise = TRUE)
}
\arguments{
\item{dob}{Date of birth.}

\item{enddate}{Date to calculate age at.}
}
\description{
For age calculations.
}
\examples{
age_calc()
}
\keyword{age}
