% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biv_olr_plot_col.R
\name{biv_olr_plot_col}
\alias{biv_olr_plot_col}
\title{For collection of datapoints for bivariate ordinal logistic regression plotting.}
\usage{
biv_olr_plot_col(meas, vars, data)
}
\arguments{
\item{meas}{outcome meassure variable name or response in data-data.frame as a string. Should be factor, preferably ordered.}

\item{vars}{variables to compare against. As vector of columnnames.}

\item{data}{dataframe of data.}
}
\description{
Use with plot_ord_odds(), model="df".
}
\keyword{olr}
