% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_overlap.R
\name{calculate_overlap}
\alias{calculate_overlap}
\title{Calculating overlap from list of identifier numbers}
\usage{
calculate_overlap(x)
}
\arguments{
\item{x}{Date of birth.}
}
\description{
Ex for creating euler diagram using the euler() function of the eulerr package.
}
\examples{

}
\keyword{overlap}
