% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie_test.R
\name{cie_test}
\alias{cie_test}
\title{A repeated regression function for change-in-estimate analysis}
\usage{
cie_test(y, v1, string, data, logistic = FALSE, cut = 0.1, v2 = NULL,
  v3 = NULL)
}
\arguments{
\item{y}{Effect meassure.}

\item{v1}{Main variable in model}

\item{string}{String of columnnames from dataframe to include. Use dput().}
}
\description{
For bivariate analyses.
}
\examples{
cie_test()
}
\keyword{change-in-estimate}
