% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_fact.R
\name{col_fact}
\alias{col_fact}
\title{Formatting multiple columns as factor}
\usage{
col_fact(string, data, levels = NULL, labels = NULL)
}
\arguments{
\item{string}{Columnnames containg strings.}

\item{data}{Dataframe}

\item{levels}{Levels for all selected columns}

\item{labels}{labels for all selected columns}
}
\description{
Depending on dply's contains()-function.
}
\keyword{factor}
