% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_num.R
\name{col_num}
\alias{col_num}
\title{Formatting multiple columns as numeric}
\usage{
col_num(string, data)
}
\arguments{
\item{string}{Columnnames containg strings.}

\item{data}{Dataframe}
}
\description{
Depending on dply's contains()-function.
}
\keyword{numeric}
