% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_olr.R
\name{comb_olr}
\alias{comb_olr}
\title{An ordinal logistic regression function for plotting}
\usage{
comb_olr(meas, vars, data)
}
\arguments{
\item{meas}{primary outcome (factor with >2 levels).}

\item{vars}{variables in model. Input as c() of columnnames, use dput().}

\item{dta}{data frame to pull variables from.}
}
\description{
Should be combined with "rep_olr()". The confint() function is rather slow, causing the whole function to hang when including many predictors and calculating the ORs with CI.
}
\examples{
comb_olr()
}
\keyword{olr}
