% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpr_check_function.R
\name{cpr_check}
\alias{cpr_check}
\title{CPR check}
\usage{
cpr_check(x)
}
\arguments{
\item{x}{cpr as "ddmmyy-xxxx".}
}
\description{
Checking validity of cpr number
}
\examples{
cpr_check("231045-0637")
}
\keyword{cpr}
