% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpr_sex_function.R
\name{cpr_sex}
\alias{cpr_sex}
\title{Determine sex from CPR}
\usage{
cpr_sex(x)
}
\arguments{
\item{x}{cpr as "ddmmyy-xxxx".}
}
\description{
Format "ddmmyy-xxxx"
}
\examples{
cpr_sex("231045-0637")
}
\keyword{cpr}
