% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_perc.R
\name{cut_perc}
\alias{cut_perc}
\title{Creates factor by percentile from numeric vector.}
\usage{
cut_perc(vec, percentile = 0.1)
}
\arguments{
\item{vec}{name of vector}

\item{percentile}{percentile for cuts. No check is made, sholud be a divisor of 1 as .25 og .1. Standard is .1 for centiles.}
}
\description{
Cuts numeric vector in equal groups set by percentile.
Export is af factor with numeric labels.
}
\examples{
  ##Tests the function by quantiles
  cut_perc(sample(1:100, 50, replace=TRUE),.25)
}
\keyword{percentile}
