% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_convert_function.R
\name{date_convert}
\alias{date_convert}
\title{Converting data format}
\usage{
date_convert(x)
}
\arguments{
\item{x}{data as as dd/mm/yyyy.}
}
\description{
Input format as dd/mm/yyyy, output is standard yyyy-mm-dd
}
\keyword{date}
