% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dob_extract_cpr_function.R
\name{dob_extract_cpr}
\alias{dob_extract_cpr}
\title{Extracting date of birth from CPR}
\usage{
dob_extract_cpr(cpr)
}
\arguments{
\item{cpr}{cpr-numbers as ddmmyy[-.]xxxx or ddmmyyxxxx. Also mixed formatting.}
}
\description{
For easy calculation.
}
\examples{
dob_extract_cpr("231045-0637")
}
\keyword{cpr}
