% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwe_allele.R
\name{hwe_allele}
\alias{hwe_allele}
\title{HWE calculation by alleles}
\usage{
hwe_allele(x, y)
}
\arguments{
\item{x}{Allele 1.}

\item{y}{Allele 2.}
}
\description{
For easy calculation.
}
\examples{
hwe_allele()
}
\keyword{hardy-weinberg-equllibrium}
