% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwe_geno.R
\name{hwe_geno}
\alias{hwe_geno}
\title{HWE calculation by genotypes}
\usage{
hwe_geno(mm, mn, nn, mo, no, oo, alleles = 2)
}
\arguments{
\item{mm}{First count of genotype.}
}
\description{
For easy calculation.
}
\examples{
hwe_geno()
}
\keyword{hardy-weinberg-equllibrium}
