% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwe_sum.R
\name{hwe_sum}
\alias{hwe_sum}
\title{Summarising function of HWE calculation}
\usage{
hwe_sum(a1, a2, f)
}
\arguments{
\item{a1}{Allele 1.}

\item{a2}{Allele 2.}

\item{f}{factor for grouping.}
}
\description{
For easy printing.
}
\examples{
hwe_sum()
}
\keyword{hardy-weinberg-equllibrium}
