% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ord_odds.R
\name{plot_ord_odds}
\alias{plot_ord_odds}
\title{Forrest plot from ordinal logistic regression}
\usage{
plot_ord_odds(x, title = NULL, dec = 3, lbls = NULL)
}
\arguments{
\item{x}{ordinal logistic regression model.}

\item{title}{plot title}

\item{dec}{decimals for labels}

\item{lbls}{labels for variable names. Carefull, as the right order is not checked automatically!}
}
\description{
Heavily inspired by https://www.r-bloggers.com/plotting-odds-ratios-aka-a-forrestplot-with-ggplot2/
}
\examples{
plot_ord_odds()
}
\keyword{forestplot}
