% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ord_odds_v2.R
\name{plot_ord_odds_v2}
\alias{plot_ord_odds_v2}
\title{Forrest plot from ordinal logistic regression, version2.}
\usage{
plot_ord_odds_v2(meas, vars, data, title = NULL, dec = 3,
  lbls = NULL, hori = "OR (95 \% CI)", vert = "Variables",
  short = FALSE, analysis = c("biv", "multi"))
}
\arguments{
\item{meas}{outcome meassure variable name or response in data-data.frame as a string. Should be factor, preferably ordered.}

\item{vars}{variables to compare against. As vector of columnnames.}

\item{data}{dataframe of data.}

\item{title}{plot title}

\item{dec}{decimals for labels}

\item{lbls}{labels for variable names. Carefull, as the right order is not checked automatically!}

\item{hori}{labels the horizontal axis (this i the y axis as the plot is rotated)}

\item{vert}{labels the horizontal axis (this i the x axis as the plot is rotated)}

\item{short}{flag to half number of ticks on horizontal axis.}

\item{analysis}{can be either "biv", or "multi", for creation of forest plot from either bivariate (unadjusted) or multivariate (adjusted) ordinal logistic regression.}
}
\description{
Heavily inspired by https://www.r-bloggers.com/plotting-odds-ratios-aka-a-forrestplot-with-ggplot2/
}
\keyword{forestplot}
