% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_diff_byvar.R
\name{print_diff_byvar}
\alias{print_diff_byvar}
\title{Printable table of three dimensional regression analysis}
\usage{
print_diff_byvar(meas, var, group, adj, data, dec = 2)
}
\arguments{
\item{meas}{outcome meassure variable name in data-data.frame as a string. Can be numeric or factor. Result is calculated accordingly.}

\item{var}{binary exposure variable to compare against (active vs placebo). As string.}

\item{group}{groups to compare, as string.}

\item{adj}{variables to adjust for, as string.}

\item{data}{dataframe of data.}

\item{dec}{decimals for results, standard is set to 2. Mean and sd is dec-1.}
}
\description{
New function ready for revision
}
\details{
Printable table of three dimensional regression analysis of group vs var for meas. By var. Includes p-values.
}
\examples{
  data('mtcars')
  mtcars$vs<-factor(mtcars$vs)
  mtcars$am<-factor(mtcars$am)
  print_diff_byvar(meas="mpg",var="vs",group = "am",adj=c("disp","wt","hp"),data=mtcars)
}
\keyword{table}
