% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_reg_diff_bin.R
\name{print_reg_diff_bin}
\alias{print_reg_diff_bin}
\title{SHOULD BE COMBINED WITH strobe_pred (try transfer the REF) - or keep as a fast/slim version??}
\usage{
print_reg_diff_bin(meas, group, var = NULL, adj, data, dec = 2)
}
\arguments{
\item{meas}{outcome measure variable name in data-data.frame as a string. Can be numeric or factor. Result is calculated accordingly.}

\item{group}{groups to compare, as string.}

\item{var}{Default is NULL.}

\item{adj}{variables to adjust for, as string.}

\item{data}{dataframe of data.}

\item{dec}{decimals for results, standard is set to 2. Mean and sd is dec-1.}
}
\description{
Print-friendly regression results for easy export to text editor
}
\details{
Printable table of regression analysis by group for outcome measure. Detects whether to perform logistic or linear regression.
output is list of
}
\examples{
  data('mtcars')
  mtcars$vs<-factor(mtcars$vs)
  mtcars$am<-factor(mtcars$am)
  print_reg_diff_bin(meas="am",group="vs",adj=c("disp","wt"),data=mtcars)
}
\keyword{strobe}
