% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_clean_csv.R
\name{redcap_clean_csv}
\alias{redcap_clean_csv}
\title{Easily format csv-files for instrument opload to REDCap}
\usage{
redcap_clean_csv(folder, name = "instrument")
}
\arguments{
\item{folder}{Path to desired folder.}

\item{name}{Desired file name.}
}
\description{
Both Numbers and Excel lacks detailed formatting options for csv-export.
This function formats csv-files for ReDCap instruments after export from said programs.
}
\examples{
csv_redcap_cleaner("/Users/user/REDCap_conversion")
}
\keyword{redcap}
