% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_biv.R
\name{rep_biv}
\alias{rep_biv}
\title{A repeated function for bivariate analyses}
\usage{
rep_biv(y, v1, string, data, method = "pval", logistic = FALSE,
  ci = FALSE, cut = 0.1, v2 = NULL, v3 = NULL)
}
\arguments{
\item{y}{Effect meassure.}

\item{v1}{Main variable in model}

\item{string}{String of columnnames to include. Obtained with dput(). Input as c().}

\item{data}{dataframe of data to analyse}

\item{method}{method of gating analysis. Can be "pval" for a simple analysis of p-values below or equal to 0.1 or "cie" for change in estimate to asses the change of the estimate when a second variable is added to the model.}
}
\description{
For bivariate analyses, for gating by p-value or change-in-estimate.
}
\examples{
rep_biv()
}
\keyword{logistic}
\keyword{regression}
