% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_glm.R
\name{rep_glm}
\alias{rep_glm}
\title{A repeated logistic regression function}
\usage{
rep_glm(y, v1, string, ci = FALSE, data, v2 = NULL, v3 = NULL)
}
\arguments{
\item{y}{Effect meassure.}

\item{v1}{Main variable in model}
}
\description{
For bivariate analyses.
}
\examples{
rep_glm()
}
\keyword{logistic}
\keyword{regression}
