% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_olr_sngl.R
\name{rep_olr_sngl}
\alias{rep_olr_sngl}
\title{A repeated ordinal logistic regression function for plotting}
\usage{
rep_olr_sngl(meas, vars, data)
}
\arguments{
\item{meas}{primary outcome (factor with >2 levels).}

\item{vars}{variables in model. Input as c() of columnnames, use dput().}

\item{dta}{data frame to pull variables from.}
}
\description{
Should be combined with "rep_olr()". For bivariate analyses. The confint() function is rather slow, causing the whole function to hang when including many predictors and calculating the ORs with CI.
}
\examples{
rep_olr_sngl()
}
\keyword{olr}
