% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strobe_diff_bygroup.R
\name{strobe_diff_bygroup}
\alias{strobe_diff_bygroup}
\title{REWRITE UNDERWAY}
\usage{
strobe_diff_bygroup(meas, var, group, adj, data, dec = 2)
}
\arguments{
\item{meas}{outcome measure variable name in data-data.frame as a string. Can be numeric or factor. Result is calculated accordingly.}

\item{var}{binary exposure variable to compare against (active vs placebo). As string.}

\item{group}{binary group to compare, as string.}

\item{adj}{variables to adjust for, as string.}

\item{data}{dataframe to subset from.}

\item{dec}{decimals for results, standard is set to 2. Mean and sd is dec-1. pval has 3 decimals.}
}
\description{
Print regression results according to STROBE
}
\details{
Printable table of two dimensional regression analysis of group vs variable for outcome measure. By group. Includes p-value
Group and variable has to be dichotomous factor.
}
\examples{
  data('mtcars')
  mtcars$vs<-factor(mtcars$vs)
  mtcars$am<-factor(mtcars$am)
  strobe_diff_bygroup(meas="mpg",var="vs",group = "am",adj=c("disp","wt"),data=mtcars)
}
\keyword{strobe}
