% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strobe_diff_twodim.R
\name{strobe_diff_twodim}
\alias{strobe_diff_twodim}
\title{Print regression results according to STROBE}
\usage{
strobe_diff_twodim(meas, group, adj, data, dec = 2)
}
\arguments{
\item{meas}{outcome meassure variable name in data-data.frame as a string. Can be numeric or factor. Result is calculated accordingly.}

\item{adj}{variables to adjust for, as string.}

\item{data}{dataframe of data.}

\item{dec}{decimals for results, standard is set to 2. Mean and sd is dec-1.}

\item{groups}{groups to compare, as string.}
}
\description{
Printable table of regression analysis by group for meas. Detects wether to perform logistic or linear regression.
}
\examples{
strobe_diff_twodim()
}
\keyword{strobe}
