% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strobe_olr.R
\name{strobe_olr}
\alias{strobe_olr}
\title{Print ordinal logistic regression results according to STROBE}
\usage{
strobe_olr(meas, vars, data, dec = 2)
}
\arguments{
\item{meas}{outcome meassure variable name in data-data.frame as a string. Can be numeric or factor. Result is calculated accordingly.}

\item{vars}{variables to compare against. As vector of columnnames.}

\item{data}{dataframe of data.}

\item{dec}{decimals for results, standard is set to 2. Mean and sd is dec-1.}
}
\description{
Printable table of ordinal logistic regression analysis oaccording to STROBE. Uses polr() funtion of the MASS-package.
}
\keyword{olr}
