% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strobe_print.R
\name{strobe_print}
\alias{strobe_print}
\title{Print regression results according to STROBE}
\usage{
strobe_print(meas, var, group, adj, data, dec = 2)
}
\arguments{
\item{meas}{outcome meassure variable name in data-data.frame as a string.}

\item{var}{exposure variable to compare against (active vs placebo). As string.}

\item{adj}{variables to adjust for, as string.}

\item{data}{dataframe of data.}

\item{dec}{decimals for results, standard is set to 2. Mean and sd is dec-1.}

\item{groups}{groups to compare, as string.}
}
\description{
Printable table of linear regression analysis of group vs var for meas.
}
\examples{
strobe_print()
}
\keyword{cpr}
