% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrRight.R
\name{substrRight}
\alias{substrRight}
\title{Substring n last characters}
\usage{
substrRight(x, n)
}
\arguments{
\item{x}{vector}

\item{n}{number of string elements to}
}
\description{
Substring n last characters
}
\examples{
  ##Kim Larsen
  cpr<-"231045-0637"
  substrRight(cpr,4)
}
\keyword{substr}
