% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_overlap.R
\name{calculate_overlap}
\alias{calculate_overlap}
\title{Calculating overlap from list of identifier numbers}
\usage{
calculate_overlap(x)
}
\arguments{
\item{x}{list of variables included. Has to be vectors of identifier numbers.}
}
\description{
To use with eulerr package for creating Euler/Venn-diagrams.
Ex for creating euler diagram using the euler() function of the eulerr package.
Up to five (5) dimensions.
}
\keyword{overlap}
