% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie_test.R
\name{cie_test}
\alias{cie_test}
\title{A repeated regression function for change-in-estimate analysis}
\usage{
cie_test(meas, vars, string, data, logistic = FALSE, cut = 0.1)
}
\arguments{
\item{string}{String of columnnames from dataframe to include. Use dput().}

\item{y}{Effect meassure.}

\item{v1}{Main variable in model}
}
\description{
For bivariate analyses. From "Modeling and variable selection in epidemiologic analysis." - S. Greenland, 1989.
}
\examples{
cie_test()
}
\keyword{change-in-estimate}
