% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie_test.R
\name{cie_test}
\alias{cie_test}
\title{A repeated regression function for change-in-estimate analysis}
\usage{
cie_test(meas, vars, string, data, logistic = FALSE, cut = 0.1)
}
\arguments{
\item{meas}{Effect meassure. Input as c() of columnnames, use dput().}

\item{vars}{variables in model. Input as c() of columnnames, use dput().}

\item{string}{variables to test. Input as c() of columnnames, use dput().}

\item{data}{data frame to pull variables from.}

\item{logistic}{flag to set logistic (TRUE) or linear (FALSE,standard) analysis.}

\item{cut}{cut value for gating if including or dropping the tested variable. As suggested bu S. Greenland (1989).}
}
\description{
For bivariate analyses. From "Modeling and variable selection in epidemiologic analysis." - S. Greenland, 1989.
}
\examples{
cie_test()
}
\keyword{change-in-estimate}
