% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpr_check_function.R
\name{cpr_check}
\alias{cpr_check}
\title{CPR check}
\usage{
cpr_check(cpr)
}
\arguments{
\item{cpr}{cpr-numbers as ddmmyy[-.]xxxx or ddmmyyxxxx. Also mixed formatting.}
}
\description{
Checking validity of cpr number
}
\examples{
cpr_check("231045-0637")
}
\keyword{cpr}
