% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dob_extract_cpr_function.R
\name{dob_extract_cpr}
\alias{dob_extract_cpr}
\title{Extracting date of birth from CPR}
\usage{
dob_extract_cpr(cpr)
}
\arguments{
\item{cpr}{cpr-numbers in format ddmmyy-xxxx.}
}
\description{
For easy calculation.
}
\examples{
dob_extract_cpr()
}
\keyword{cpr}
