% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_plot.R
\name{euler_plot}
\alias{euler_plot}
\title{Creates Euler model from list of identifier numbers.}
\usage{
euler_plot(x, shape)
}
\arguments{
\item{x}{list of variables included. Has to be vectors of identifier numbers.}

\item{shape}{same as for euler(). These includes c("circle","ellipse").}
}
\description{
Calculates overlaps and uses eulerr package to create Euler/Venn-diagrams. Use plot() to create diagram.
Combined with an evolved calculate.overlap() from the VennDiagram library.
Up to five (5) dimensions. Limit set by the complexity of combinations. euler() supports more.
}
\keyword{overlap}
