% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_lm.R
\name{rep_lm}
\alias{rep_lm}
\title{A repeated linear regression function}
\usage{
rep_lm(meas, vars = NULL, string, ci = FALSE, data, fixed.var = FALSE)
}
\arguments{
\item{meas}{Effect meassure. Input as c() of columnnames, use dput().}

\item{vars}{variables in model. Input as c() of columnnames, use dput().}

\item{string}{variables to test. Input as c() of columnnames, use dput().}

\item{ci}{flag to get results as OR with 95 percent confidence interval.}

\item{data}{data frame to pull variables from.}

\item{fixed.var}{flag to set "vars" as fixed in the model. When FALSE, then true bivariate linear regression is performed.}
}
\description{
For bivariate analyses, to determine which variables to include in adjusted model.
}
\keyword{linear}
\keyword{regression}
