% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strobe_diff_byvar.R
\name{strobe_diff_byvar}
\alias{strobe_diff_byvar}
\title{Print regression results according to STROBE}
\usage{
strobe_diff_byvar(meas, var, group, adj, data, dec = 2)
}
\arguments{
\item{meas}{outcome meassure variable name in data-data.frame as a string. Can be numeric or factor. Result is calculated accordingly.}

\item{var}{binary exposure variable to compare against (active vs placebo). As string.}

\item{adj}{variables to adjust for, as string.}

\item{data}{dataframe of data.}

\item{dec}{decimals for results, standard is set to 2. Mean and sd is dec-1.}

\item{groups}{groups to compare, as string.}
}
\description{
Printable table of three dimensional regression analysis of group vs var for meas. By var. Includes p-values.
}
\examples{
strobe_diff_byvar()
}
\keyword{strobe}
