% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strobe_olr.R
\name{strobe_olr}
\alias{strobe_olr}
\title{Print regression results according to STROBE}
\usage{
strobe_olr(meas, var, adj, data, dec = 2)
}
\arguments{
\item{meas}{outcome meassure variable name in data-data.frame as a string. Can be numeric or factor. Result is calculated accordingly.}

\item{var}{exposure variable to compare against (active vs placebo). As string.}

\item{adj}{variables to adjust for, as string.}

\item{data}{dataframe of data.}

\item{dec}{decimals for results, standard is set to 2. Mean and sd is dec-1.}
}
\description{
Printable table of linear regression analysis of group vs var for meas. By group.
}
\keyword{olr}
\keyword{strobe}
