% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strobe_pred.R
\name{strobe_pred}
\alias{strobe_pred}
\title{Regression model of predictors according to STROBE, bi- and multivariate.}
\usage{
strobe_pred(meas, adj, data, dec = 2, n.by.adj = FALSE)
}
\arguments{
\item{meas}{binary outcome meassure variable, column name in data.frame as a string. Can be numeric or factor. Result is calculated accordingly.}

\item{adj}{variables to adjust for, as string.}

\item{data}{dataframe of data.}

\item{dec}{decimals for results, standard is set to 2. Mean and sd is dec-1.}

\item{n.by.adj}{flag to indicate wether to count number of patients in adjusted model or overall for outcome meassure not NA.}
}
\description{
Printable table of regression model according to STROBE. Includes borth bivariate and multivariate in the same table. Output is a list, with the first item being the main "output" as a dataframe. Automatically uses logistic regression model for dichotomous outcome variable and linear regression model for continous outcome variable.
}
\keyword{logistic}
