% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_pred.R, R/strobe_pred.R
\name{strobe_pred}
\alias{strobe_pred}
\title{OBSOLETE - use 'print_pred'}
\usage{
strobe_pred(meas, adj, data, dec = 2, n.by.adj = FALSE, p.val = FALSE)

strobe_pred(meas, adj, data, dec = 2, n.by.adj = FALSE, p.val = FALSE)
}
\arguments{
\item{meas}{binary outcome measure variable, column name in data.frame as a string. Can be numeric or factor. Result is calculated accordingly.}

\item{adj}{variables to adjust for, as string.}

\item{data}{dataframe of data.}

\item{dec}{decimals for results, standard is set to 2. Mean and sd is dec-1.}

\item{n.by.adj}{flag to indicate whether to count number of patients in adjusted model or overall for outcome measure not NA.}

\item{p.val}{flag to include p-values in table, set to FALSE as standard.}
}
\description{
Regression model of predictors according to STROBE, bi- and multivariable.
}
\details{
Printable table of regression model according to STROBE for linear or binary outcome-variables.
Includes both bivariate and multivariate in the same table.
Output is a list, with the first item being the main "output" as a dataframe.
Automatically uses logistic regression model for dichotomous outcome variable and linear regression model for continuous outcome variable. Linear regression will give estimated adjusted true mean in list.
For logistic regression gives count of outcome variable pr variable level.

Regression model of predictors according to STROBE, bi- and multivariable.
}
\details{
Printable table of regression model according to STROBE for linear or binary outcome-variables.
Includes both bivariate and multivariate in the same table.
Output is a list, with the first item being the main "output" as a dataframe.
Automatically uses logistic regression model for dichotomous outcome variable and linear regression model for continuous outcome variable. Linear regression will give estimated adjusted true mean in list.
For logistic regression gives count of outcome variable pr variable level.
}
\keyword{logistic}
